requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
ta>=0.7.0
alpaca-trade-api>=3.0.0
python-dotenv>=0.19.0
pytz>=2021.1
joblib>=1.0.1
matplotlib>=3.4.2
seaborn>=0.11.1
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
SpeechRecognition==3.10.0 